requirements.txt
Flask>=3.0.0
Flask-WTF>=1.2.1
python-dotenv>=1.0.1
requests>=2.31.0
stripe>=6.0.0
